refactor(step1): Input, Output 패키지 변경 -> view
